Checkout

+++ database/migrations/2020_08_13_224515_create_orders_table.php
@@ -20,7 +20,7 @@ class CreateOrdersTable extends Migration
             $table->string('city');
             $table->string('state');
             $table->string('zip');
-            $table->integer('amount');
+            $table->float('amount');
             $table->id();
             $table->timestamps();
         });
